Fix combined graph starts and ends.
Previously, when combining graphs, we included everything from max(first)
to max(last).  The reason was that a relay that has not been running for
weeks shouldn't be included in a graph covering the past 3 days.

This approach did not take into account that a relay could have first
appeared a few hours ago, and then the combined graph would have shown
only those few hours.

The better approach is to include everything from 3 days ago to now, or
whatever time period we're combining.

Problem spotted by Nikita Borisov.